Windows package build fix.
git-svn-id: https://zxtune.googlecode.com/svn/trunk@1155 e19a5227-2556-0410-816f-a35683127991

+++ src/core/plugins/containers/raw_supp.cpp
@@ -61,7 +61,7 @@ namespace
         0 == str.find(PREFIX) &&
         Parameters::ConvertFromString(str.substr(PREFIX.size()), res))
     {
-      offset = res;
+      offset = static_cast<std::size_t>(res);
       return true;
     }
     return false;
